BAS-73: the component registry needs no sampler
CI leaves pymc and arviz out on purpose, and the registry that names a
component, validates it and reads its columns lived in the module that
imports both. Lifted into src/models/pa_components.py, re-exported from
pa_rate, and the eval harness resolves components from the light module.
